+++ lms/djangoapps/bulk_email/tests/test_course_optout.py
@@ -38,6 +38,12 @@ class TestOptoutCourseEmails(ModuleStoreTestCase):
 
         self.client.login(username=self.student.username, password="test")
 
+        self.send_mail_url = reverse('send_email', kwargs={'course_id': self.course.id})
+        self.success_content = {
+            'course_id': self.course.id,
+            'success': True,
+        }
+
     def tearDown(self):
         """
         Undo all patches.
@@ -47,10 +53,17 @@ class TestOptoutCourseEmails(ModuleStoreTestCase):
     def navigate_to_email_view(self):
         """Navigate to the instructor dash's email view"""
         # Pull up email view on instructor dashboard
+<<<<<<< HEAD
         url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id.to_deprecated_string()})
+=======
+        url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id})
+        # Response loads the whole instructor dashboard, so no need to explicitly
+        # navigate to a particular email section
+>>>>>>> edx/master
         response = self.client.get(url)
-        email_link = '<a href="#" onclick="goto(\'Email\')" class="None">Email</a>'
+        email_section = '<div class="vert-left send-email" id="section-send-email">'
         # If this fails, it is likely because ENABLE_INSTRUCTOR_EMAIL is set to False
+<<<<<<< HEAD
         self.assertTrue(email_link in response.content)
 
         # Select the Email view of the instructor dash
@@ -60,6 +73,9 @@ class TestOptoutCourseEmails(ModuleStoreTestCase):
         response = self.client.get(url)
         selected_email_link = '<a href="#" onclick="goto(\'Email\')" class="selectedmode">Email</a>'
         self.assertTrue(selected_email_link in response.content)
+=======
+        self.assertTrue(email_section in response.content)
+>>>>>>> edx/master
 
     @patch.dict(settings.FEATURES, {'ENABLE_INSTRUCTOR_EMAIL': True, 'REQUIRE_COURSE_EMAIL_AUTH': False})
     def test_optout_course(self):
@@ -77,15 +93,18 @@ class TestOptoutCourseEmails(ModuleStoreTestCase):
         self.client.login(username=self.instructor.username, password="test")
         self.navigate_to_email_view()
 
+<<<<<<< HEAD
         url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id.to_deprecated_string()})
+=======
+>>>>>>> edx/master
         test_email = {
             'action': 'Send email',
-            'to_option': 'all',
+            'send_to': 'all',
             'subject': 'test subject for all',
             'message': 'test message for all'
         }
-        response = self.client.post(url, test_email)
-        self.assertContains(response, "Your email was successfully queued for sending.")
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         # Assert that self.student.email not in mail.to, outbox should be empty
         self.assertEqual(len(mail.outbox), 0)
@@ -106,16 +125,18 @@ class TestOptoutCourseEmails(ModuleStoreTestCase):
         self.client.login(username=self.instructor.username, password="test")
         self.navigate_to_email_view()
 
+<<<<<<< HEAD
         url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id.to_deprecated_string()})
+=======
+>>>>>>> edx/master
         test_email = {
             'action': 'Send email',
-            'to_option': 'all',
+            'send_to': 'all',
             'subject': 'test subject for all',
             'message': 'test message for all'
         }
-        response = self.client.post(url, test_email)
-
-        self.assertContains(response, "Your email was successfully queued for sending.")
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         # Assert that self.student.email in mail.to
         self.assertEqual(len(mail.outbox), 1)