+++ lms/djangoapps/bulk_email/tests/test_err_handling.py
@@ -2,6 +2,8 @@
 """
 Unit tests for handling email sending errors
 """
+import json
+
 from itertools import cycle
 from mock import patch
 from smtplib import SMTPDataError, SMTPServerDisconnected, SMTPConnectError
@@ -53,7 +55,16 @@ class TestEmailErrors(ModuleStoreTestCase):
 
         # load initial content (since we don't run migrations as part of tests):
         call_command("loaddata", "course_email_template.json")
+<<<<<<< HEAD
         self.url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id.to_deprecated_string()})
+=======
+        self.url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id})
+        self.send_mail_url = reverse('send_email', kwargs={'course_id': self.course.id})
+        self.success_content = {
+            'course_id': self.course.id,
+            'success': True,
+        }
+>>>>>>> edx/master
 
     def tearDown(self):
         patch.stopall()
@@ -67,15 +78,16 @@ class TestEmailErrors(ModuleStoreTestCase):
         get_conn.return_value.send_messages.side_effect = SMTPDataError(455, "Throttling: Sending rate exceeded")
         test_email = {
             'action': 'Send email',
-            'to_option': 'myself',
+            'send_to': 'myself',
             'subject': 'test subject for myself',
             'message': 'test message for myself'
         }
-        self.client.post(self.url, test_email)
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         # Test that we retry upon hitting a 4xx error
         self.assertTrue(retry.called)
-        (_, kwargs) = retry.call_args
+        (__, kwargs) = retry.call_args
         exc = kwargs['exc']
         self.assertIsInstance(exc, SMTPDataError)
 
@@ -95,11 +107,12 @@ class TestEmailErrors(ModuleStoreTestCase):
 
         test_email = {
             'action': 'Send email',
-            'to_option': 'all',
+            'send_to': 'all',
             'subject': 'test subject for all',
             'message': 'test message for all'
         }
-        self.client.post(self.url, test_email)
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         # We shouldn't retry when hitting a 5xx error
         self.assertFalse(retry.called)
@@ -119,14 +132,15 @@ class TestEmailErrors(ModuleStoreTestCase):
         get_conn.return_value.open.side_effect = SMTPServerDisconnected(425, "Disconnecting")
         test_email = {
             'action': 'Send email',
-            'to_option': 'myself',
+            'send_to': 'myself',
             'subject': 'test subject for myself',
             'message': 'test message for myself'
         }
-        self.client.post(self.url, test_email)
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         self.assertTrue(retry.called)
-        (_, kwargs) = retry.call_args
+        (__, kwargs) = retry.call_args
         exc = kwargs['exc']
         self.assertIsInstance(exc, SMTPServerDisconnected)
 
@@ -140,14 +154,15 @@ class TestEmailErrors(ModuleStoreTestCase):
 
         test_email = {
             'action': 'Send email',
-            'to_option': 'myself',
+            'send_to': 'myself',
             'subject': 'test subject for myself',
             'message': 'test message for myself'
         }
-        self.client.post(self.url, test_email)
+        response = self.client.post(self.send_mail_url, test_email)
+        self.assertEquals(json.loads(response.content), self.success_content)
 
         self.assertTrue(retry.called)
-        (_, kwargs) = retry.call_args
+        (__, kwargs) = retry.call_args
         exc = kwargs['exc']
         self.assertIsInstance(exc, SMTPConnectError)
 
@@ -163,7 +178,7 @@ class TestEmailErrors(ModuleStoreTestCase):
         task_input = {"email_id": -1}
         with self.assertRaises(CourseEmail.DoesNotExist):
             perform_delegate_email_batches(entry.id, course_id, task_input, "action_name")  # pylint: disable=E1101
-        ((log_str, _, email_id), _) = mock_log.warning.call_args
+        ((log_str, __, email_id), __) = mock_log.warning.call_args
         self.assertTrue(mock_log.warning.called)
         self.assertIn('Failed to get CourseEmail with id', log_str)
         self.assertEqual(email_id, -1)