+++ lms/djangoapps/instructor/tests/test_email.py
@@ -35,7 +35,11 @@ class TestNewInstructorDashboardEmailViewMongoBacked(ModuleStoreTestCase):
         self.client.login(username=instructor.username, password="test")
 
         # URL for instructor dash
+<<<<<<< HEAD
         self.url = reverse('instructor_dashboard_2', kwargs={'course_id': self.course.id.to_deprecated_string()})
+=======
+        self.url = reverse('instructor_dashboard', kwargs={'course_id': self.course.id})
+>>>>>>> edx/master
         # URL for email view
         self.email_link = '<a href="" data-section="send_email">Email</a>'
 
@@ -123,7 +127,11 @@ class TestNewInstructorDashboardEmailViewXMLBacked(ModuleStoreTestCase):
         self.client.login(username=instructor.username, password="test")
 
         # URL for instructor dash
+<<<<<<< HEAD
         self.url = reverse('instructor_dashboard_2', kwargs={'course_id': self.course_key.to_deprecated_string()})
+=======
+        self.url = reverse('instructor_dashboard', kwargs={'course_id': self.course_name})
+>>>>>>> edx/master
         # URL for email view
         self.email_link = '<a href="" data-section="send_email">Email</a>'