+++ lms/djangoapps/instructor/tests/test_legacy_anon_csv.py
@@ -53,7 +53,11 @@ class TestInstructorDashboardAnonCSV(ModuleStoreTestCase, LoginEnrollmentTestCas
 
     def test_download_anon_csv(self):
         course = self.toy
+<<<<<<< HEAD
         url = reverse('instructor_dashboard', kwargs={'course_id': course.id.to_deprecated_string()})
+=======
+        url = reverse('instructor_dashboard_legacy', kwargs={'course_id': course.id})
+>>>>>>> edx/master
 
         with patch('instructor.views.legacy.unique_id_for_user') as mock_unique:
             mock_unique.return_value = 42