+++ lms/djangoapps/instructor/tests/test_legacy_gradebook.py
@@ -68,10 +68,14 @@ class TestGradebook(ModuleStoreTestCase):
                     module_state_key=item.location
                 )
 
+<<<<<<< HEAD
         self.response = self.client.get(reverse(
             'gradebook',
             args=(self.course.id.to_deprecated_string(),)
         ))
+=======
+        self.response = self.client.get(reverse('gradebook_legacy', args=(self.course.id,)))
+>>>>>>> edx/master
 
     def test_response_code(self):
         self.assertEquals(self.response.status_code, 200)