+++ docs/en_us/course_authors/source/change_log.rst
@@ -12,6 +12,9 @@ May, 2014
 
    * - Date
      - Change
+     - 05/13/14
+     - Updated the :ref:`Enrollment` section to reflect that usernames or email
+       addresses can be used to batch enroll students.
    * - 05/09/14
      - Updated :ref:`Assigning_discussion_roles` with a note about course staff requiring
        explicit granting of discussion administration roles.